# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: BileMo-API
  version: '1'
servers:
  -
    url: 'https://127.0.0.1:8000'
    description: 'Dev Server'
paths:
...
  /api/reseller:
    get:
      tags:
        - Reseller
      summary: 'Get reseller detail'
      description: 'Get reseller detail'
      operationId: 'App\Controller\ResellerController::resellerInfo'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResellerDetail'
              example:
                id: 1
                email: test@email.fr
                password: test
                _links:
                  self: { href: 'https://127.0.0.1:8000/api/reseller' }
                  delete: { href: 'https://127.0.0.1:8000/api/reseller' }
                  create_user: { href: 'https://127.0.0.1:8000/api/user' }
        '401':
          description: 'Bearer token missing'
          content:
            application/json:
              schema: {  }
              example:
                code: '401'
                message: 'JWT Token not found'
      security:
        -
          bearer: []
    delete:
      tags:
        - Reseller
      summary: 'Delete reseller'
      description: 'Delete reseller'
      operationId: 'App\Controller\ResellerController::resellerDelete'
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                format: string
              example:
                - ''
        '401':
          description: 'Bearer token missing'
          content:
            application/json:
              schema: {  }
              example:
                code: '401'
                message: 'JWT Token not found'
      security:
        -
          bearer: []
  /create/reseller:
    post:
      tags:
        - Reseller
      summary: 'Create reseller'
      description: 'Create reseller'
      operationId: 'App\Controller\ResellerController::resellerCreate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  description: 'User email'
                  type: string
                password:
                  description: 'User passord'
                  type: string
              type: object
              example:
                email: test@email.fr
                password: test
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResellerDetail'
              example:
                id: 1
                email: test@email.fr
                password: test
                _links:
                  self: { href: 'https://127.0.0.1:8000/api/reseller' }
                  delete: { href: 'https://127.0.0.1:8000/api/reseller' }
                  create_user: { href: 'https://127.0.0.1:8000/api/user' }
        '400':
          description: 'Invalid input, or duplicate Reseller'
          content:
            application/json:
              schema: {  }
              example:
                message: 'Bad request. Check your parameters'
      security:
        -
          bearer: []
...
components:
  schemas:
...
    ResellerDetail:
      description: ResellerDetail
      properties:
        id:
          type: integer
        email:
          type: string
        password:
          type: string
      type: object
...
  securitySchemes:
    bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer